Set locale after reading ~/.rcrc
rc does not handle umlauts correctly, if locale environment variables are first
set in ~/.rcrc. Simply calling setlocale(LC_CTYPE, "") after reading ~/.rcrc is
not enough, because the environment is only updated directly before a call to
exec(3). The new utility function applylocale() does the same as glibc
setlocale(LC_CTYPE, "") but it looks up the not yet exported variables LC_ALL,
LC_CTYPE and LANG, using the value of the first defined one instead of "".