Completion of command names and variables
New completion entry functions for readline specific code:
  - compl_var for variables
  - compl_fn for functions
  - compl_builtin for builtins
  - compl_extcmd for external commands in path
  - compl_command, which combines the last three

The first three are simple wrappers for compl_name, which iterates over a
typical rc table (an array of structs) yielding the name fields.

The appropriate completion entry function is chosen based on the
character/token that preceeds the word to complete: '$' for variable,
beginning of line or one of [`@|&(){;] for command. Otherwise the shell falls
back to readline's default completion, which it also does, if the chosen
completion entry function finds no match.