Custom readline commands for explicit completions
New commands:
  - complete-command, per default bound to "M-!"
  - complete-filename, "M-/"
  - complete-variable, "M-$"

These commands are useful, if one wants a specific type of completion which is
not performed by the context-senstive default one, e.g. filename completion at
the beginning of the line or command completion for the second word (after
sudo).

The readline code now also sets the application name to "rc", making it
possible to specify different bindings for these commands in ~/.inputrc:
  $if rc
  "\C-x\C-[": complete-command
  "\C-[\C-x": complete-filename
  "\C-x\C-v": complete-variable
  $endif