Tilde expansion
The expansion is performed in glob(). It is only possible after glom(), because
the free-caret in "~user" happens there. In order to differentiate unquoted
tilde characters from quoted ones this long after parsing, especially after the
conversion of nWord to List by word() in glom(), the tilde character has to
become a matching character like *, ?, [. Only one additional check is required
in match() to prevent unwanted effects of this change.