Parse equals sign as keyword
The parser treats unexpected keyword tokens (eg. if, while, ~) as literals and
not as syntax errors. It does not do this for the equals sign, but it could, as
it only has a special meaning at the beginning of a command, much like the
keywords. Doing this allows one to write commands like this:

    dd if'='/dev/zero of'='data bs'='1024 count'='1

without the quotes around the equals signs. All shells from the Bourne family
do this.

For this to work, the lexer must not automatically terminate words at equals
signs, instead it has to generate a free caret around it like for the twiddle
and other such special characters.

In the parser's grammar we need to get rid of the left recursion in the rules
for 'simple' to turn the reduce/reduce conflict at 'first' on '=' (empty ->
'optcaret' or 'first' -> 'simple') into a shift/reduce conflict, which then
can be solved with precedence/associativity so that the reduction of the
empty 'optcaret' wins over shifting '='.